% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Honolulu_White.R
\name{Honolulu_White}
\alias{Honolulu_White}
\title{A Honolulu_White Function}
\usage{
Honolulu_White(new = TRUE)
}
\arguments{
\item{new}{Are you a new guy? Defaults to TRUE.}
}
\description{
This function was written by a person with little experieces in R package generating from Honolulu.
}
\examples{
Honolulu_White()
}
\keyword{new}
